UPDATED STAFF & ADMIN AND ALSO STAFF AND ADMIN PROFILE
Too many files been added, not sure if the inc needed. I just add it for future reference, if dont want can just delete it. If keeping it, it can be quite convenient as we don't have to repeat the same code but we have to slowly edit the page one by one to organize it